refactor: auth랑 member 관련 기능 분기
