+++ backend/src/main/java/org/example/be/security/config/SecurityConfig.java
@@ -49,8 +49,7 @@ public class SecurityConfig {
 		http
 			.authorizeHttpRequests(auth -> auth
 				.requestMatchers(HttpMethod.OPTIONS, "/**").permitAll()
-				.requestMatchers(HttpMethod.POST, "/members").permitAll()
-				.requestMatchers(HttpMethod.POST, "/members/login").permitAll()
+				.requestMatchers("/auth/**").permitAll()
 				.requestMatchers("/oauth2/**").permitAll()
 				.requestMatchers("/favicon.ico").permitAll()
 				.requestMatchers("/.well-known/**").permitAll()
